backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
pydantic
face_recognition
python-dotenv
